% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artscore.R
\name{Artscore}
\alias{Artscore}
\title{Get an Artscore from a list of species and a specific habitat type}
\usage{
Artscore(
  ScientificName = NULL,
  Common_name = NULL,
  Habitat_name = NULL,
  Habitat_code = NULL
)
}
\arguments{
\item{ScientificName}{a vector with the list of found species in a location the
full list can be found in the \code{\link{Species_List}} dataset in the
Scientific_name column.}

\item{Common_name}{Only use if you didn't provide the Scientific name,
a vector with the list of found species in a location the
full list can be found in the \code{\link{Species_List}} dataset in the
Danish_name column.}

\item{Habitat_name}{a character with the habitat found in the location,
a full list can be found in the \code{\link{Habitat_List}} dataset in the
habitat_name column.}

\item{Habitat_code}{Only use if you didn't provide the Habitat_name,
a character with the habitat found in the location,  a full list can be
found in the \code{\link{Habitat_List}} dataset in the  habitat_name
column.}
}
\value{
a data frame with the intermediate calculations and the Artscore
}
\description{
Get an Artscore from a list of species and a specific habitat type
}
\details{
The dataframe returns the following calculated values

m = Adjusted average score for the test field.

a(b) = Number (a) contribution species (b)
(number with species score above 0). Coded as a_b in the data
frame.

s = Artsum calculated as follows:

\deqn{s = m \times a(b)}

m(a) = Average adjusted mean score, coded as m_a in the data
frame

A(s) = Artcore index, coded as A_s in the data frame calculated
as follows
}
