% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWeights.R
\name{GetWeights}
\alias{GetWeights}
\title{Get the weights and subweights for a particular habitat}
\usage{
GetWeights(Habitat_name = NULL, Habitat_code = NULL)
}
\arguments{
\item{Habitat_name}{a character with the habitat found in the location,
a full list can be found in the \code{\link{Habitat_List}} dataset in the
habitat_name column.}

\item{Habitat_code}{Only use if you didn't provide the Habitat_name,
a character with the habitat found in the location,  a full list can be
found in the \code{\link{Habitat_List}} dataset in the  habitat_name
column.}
}
\value{
a data frame with the weights and subweights of each habitat
}
\description{
Get the weights and subweights for a particular habitat
}
\examples{

data("Habitat_List")

GetWeights(Habitat_name = Habitat_List$habitat_name[10])

}
