% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Habitat_List.R
\docType{data}
\name{Habitat_List}
\alias{Habitat_List}
\title{List of Habitats available to calculate Artscores.}
\format{
A data frame with 33 rows and 2 variables:
\describe{
\item{Code}{Numeric code of the habitat}
\item{habitat_name}{Danish name of the habitat}
...
}
}
\usage{
Habitat_List
}
\description{
A dataset containing the danish names and codes of 33
habitat types.
}
\keyword{datasets}
