% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naturtilstandindeks.R
\name{Naturtilstandindeks}
\alias{Naturtilstandindeks}
\title{Natural condition index}
\usage{
Naturtilstandindeks(Artscore, Strukturindeks)
}
\arguments{
\item{Artscore}{a numeric value, can be calculated using the,
\code{\link{Artscore}} function.}

\item{Strukturindeks}{a numeric value, can be calculated using the,
\code{\link{Strukturindeks}} function.}
}
\value{
a numeric value with the Natual condition index
}
\description{
Natural condition index
}
\examples{

data("Habitat_List")
data("Species_List")

set.seed(2022)

Ascore <- Artscore(Common_name = sample(Species_List$Danish_name, 10),
Habitat_code = Habitat_List$Code[10])
Naturtilstandindeks(Artscore = Ascore$Artsindex,
Strukturindeks = 0.6)

}
