% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Species_List.R
\docType{data}
\name{Species_List}
\alias{Species_List}
\title{List of species available to calculate Artscores.}
\format{
A data frame with 1636 rows and 2 variables:
\describe{
\item{Scientific_name}{Scientific name of the species}
\item{Danish_name}{Danish name of the species}
...
}
}
\usage{
Species_List
}
\description{
A dataset containing the danish common names and scientific names of 1,636
species.
}
\keyword{datasets}
