% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strukturindeks.R
\name{Strukturindeks}
\alias{Strukturindeks}
\title{Get an Struktureindeks value from a specific habitat type}
\usage{
Strukturindeks(Habitat_name = NULL, Habitat_code = NULL)
}
\arguments{
\item{Habitat_name}{a character with the habitat found in the location,
a full list can be found in the \code{\link{Habitat_List}} dataset in the
habitat_name column.}

\item{Habitat_code}{Only use if you didn't provide the Habitat_name,
a character with the habitat found in the location,  a full list can be
found in the \code{\link{Habitat_List}} dataset in the  habitat_name
column.}
}
\value{
an Struktureindeks value
}
\description{
Get an Struktureindeks value from a specific habitat type
}
\examples{
\dontrun{
data("Habitat_List")

set.seed(2022)

Strukturindeks(Habitat_name = Habitat_List$habitat_name[10])
}



}
