% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Habitat_List.R, R/Species_List.R
\docType{data}
\name{Species_List}
\alias{Species_List}
\title{List of Habitats available to calculate Artscores.}
\format{
A data frame with 33 rows and 2 variables:
\describe{
\item{Code}{Numeric code of the habitat}
\item{habitat_name}{Danish name of the habitat}
...
}

A data frame with 1636 rows and 2 variables:
\describe{
\item{Scientific_name}{Scientific name of the species}
\item{Danish_name}{Danish name of the species}
...
}
}
\usage{
Species_List

Species_List
}
\description{
A dataset containing the danish names and codes of 33
habitat types.

A dataset containing the danish common names and scientific names of 1,636
species.
}
\keyword{datasets}
